requirements.txt
pandas
python-dotenv
requests
beautifulsoup4
forex_python
numpy
pytest
